refactor metrics.
